don't call validators on keys of dictionaries, fix #254
