+++ pydantic/__init__.py
@@ -1,9 +1,10 @@
 # flake8: noqa
+from .class_validators import validator
 from .env_settings import BaseSettings
 from .error_wrappers import ValidationError
 from .errors import *
 from .fields import Required
-from .main import BaseConfig, BaseModel, create_model, validate_model, validator
+from .main import BaseConfig, BaseModel, create_model, validate_model
 from .parse import Protocol
 from .types import *
 from .version import VERSION
