+++ pydantic/version.py
@@ -2,4 +2,4 @@ from distutils.version import StrictVersion
 
 __all__ = ['VERSION']
 
-VERSION = StrictVersion('0.17.1a1')
+VERSION = StrictVersion('0.18a1')
